Move key combination parser/unparser to core/keyboard
